gracefully shutdown test services

+++ test-utils/src/manager.rs
@@ -81,6 +81,20 @@ impl Manager {
 
     pub fn kill(&mut self) {
         if let Some(mut process) = self.process.take() {
+            println!("Stopping: {}", self.instance_name);
+            if let Ok(Some(_)) = process.try_wait() {
+                return;
+            }
+            let _ = Command::new("kill").arg(process.id().to_string()).output();
+            let mut count = 0;
+            while count < 30 {
+                if let Ok(Some(_)) = process.try_wait() {
+                    println!("{} stopped after {count} secs", self.instance_name);
+                    return;
+                }
+                std::thread::sleep(Duration::from_secs(1));
+                count += 1;
+            }
             println!("Killing: {}", self.instance_name);
             process.kill().unwrap_or_default();
             process.wait().unwrap();
